Move file in correct folder
