+++ app/src/androidTest/java/com/github/fribourgsdp/radio/ui/KeepOrDismissPlaylistDialogTest.kt
@@ -1,17 +1,19 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.ui
 
 import androidx.fragment.app.testing.launchFragment
 import androidx.test.espresso.Espresso
 import androidx.test.espresso.action.ViewActions
 import androidx.test.espresso.matcher.RootMatchers
 import androidx.test.espresso.matcher.ViewMatchers
+import com.github.fribourgsdp.radio.data.view.KeepOrDismissPlaylistDialog
+import com.github.fribourgsdp.radio.R
 import org.junit.Assert
 import org.junit.Test
 
 class KeepOrDismissPlaylistDialogTest {
 
     private var picked = KeepOrDismissPlaylistDialog.Choice.KEEP
-    private val testListener = object: KeepOrDismissPlaylistDialog.OnPickListener{
+    private val testListener = object: KeepOrDismissPlaylistDialog.OnPickListener {
         override fun onPick(choice: KeepOrDismissPlaylistDialog.Choice){
             picked = choice
         }
@@ -22,7 +24,7 @@ class KeepOrDismissPlaylistDialogTest {
         picked = KeepOrDismissPlaylistDialog.Choice.DISMISS
         with(launchFragment (
             themeResId = R.style.Theme_Radio,
-            instantiate = { KeepOrDismissPlaylistDialog(testListener)}
+            instantiate = { KeepOrDismissPlaylistDialog(testListener) }
         )){
             Espresso.onView(ViewMatchers.withId(R.id.keepPlaylistButton))
                 .inRoot(RootMatchers.isDialog())
@@ -36,7 +38,7 @@ class KeepOrDismissPlaylistDialogTest {
         picked = KeepOrDismissPlaylistDialog.Choice.KEEP
         with(launchFragment (
             themeResId = R.style.Theme_Radio,
-            instantiate = { KeepOrDismissPlaylistDialog(testListener)}
+            instantiate = { KeepOrDismissPlaylistDialog(testListener) }
         )){
             Espresso.onView(ViewMatchers.withId(R.id.dismissPlaylistButton))
                 .inRoot(RootMatchers.isDialog())
