+++ app/src/main/java/com/github/fribourgsdp/radio/data/view/KeepOrDismissPlaylistDialog.kt
@@ -1,4 +1,4 @@
-package com.github.fribourgsdp.radio
+package com.github.fribourgsdp.radio.data.view
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Button
 import androidx.fragment.app.DialogFragment
+import com.github.fribourgsdp.radio.R
 
 class KeepOrDismissPlaylistDialog (private val listener: OnPickListener) : DialogFragment() {
